Move creating store into separate file Signed-off-by: Raimund Schlüßler <raimund.schluessler@mailbox.org>
